1000000000l default fps has been broken on ntsc for a LONG time because of this nonsense! (29.00?!?) no idea why no one caught it, i just have fps=29.97 in my config file... i hope the bsd code works; i can't test it. btw, the (int)*(void **)arg stuff is total nonsense, no idea how that even works...
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@10020 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/tv.c
@@ -449,7 +449,7 @@ int demux_open_tv(demuxer_t *demuxer)
 
     if(!sh_video->fps)
     {
-        int tmp;
+        float tmp;
         if (funcs->control(tvh->priv, TVI_CONTROL_VID_GET_FPS, &tmp) != TVI_CONTROL_TRUE)
              sh_video->fps = 25.0f; /* on PAL */
         else sh_video->fps = tmp;
